% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densdist.R
\name{densdist}
\alias{densdist}
\title{Density and Distance}
\usage{
densdist(RCAmat, PROXmat, measure = c("density", "distance"))
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}

\item{PROXmat}{A product-by-product matrix with the proximity values.}

\item{measure}{A character string specifying the measure to compute:
\enumerate{
\item "density": the new product’s proximity to the economy's current export
basket.
\item "distance": the economy's ability to enter a specific product.
}}
}
\value{
A economy-by-product matrix with the measured values.
}
\description{
Compute the density and distance \insertCite{hausmann2006structural}{pecan}
based on the RCA and proximity values.
}
\references{
\insertAllCited{}
}
