% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrep.R
\name{discrep}
\alias{discrep}
\title{discrep}
\usage{
discrep(path, initial1, initial2, aidfrom, aidto)
}
\arguments{
\item{path}{folder in which the two coders' method extraction sheets in .csv format are}

\item{initial1}{initials for first coder in quotes}

\item{initial2}{initials for second coder in quotes}

\item{aidfrom}{article id from in quotes}

\item{aidto}{article id to in quotes}
}
\description{
-micah project exclusive- compute discrepancies between two coders
}
\examples{
discrep(path = "/Users/phoebelam/Desktop/micah", initial1 = "kk", initial2 = "arp", aidfrom = "b21", aidto = "b26")

}
