% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divubq.R
\name{divubq}
\alias{divubq}
\title{Diversity and Ubiquity}
\usage{
divubq(RCAmat, measure = c("diversity", "ubiquity"))
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}

\item{measure}{A character string specifying the measure to compute:
\enumerate{
\item "diversity": the number of products that an economy produces.
\item "ubiquity": the number of economies that export that product.
}}
}
\value{
A numeric vector with the measured values.
}
\description{
Compute the diversity/diversification and ubiquity
\insertCite{hidalgo2009dynamics}{pecan} based on the RCA values.
}
\references{
\insertAllCited{}
}
