% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excl.R
\name{excl}
\alias{excl}
\title{exclude}
\usage{
excl(x, sd)
}
\arguments{
\item{x}{vector of data}

\item{sd}{the number of standard deviations from mean as the criterion for excluding}
}
\description{
exclude based on user input sd from the mean
}
