% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluemoon.R
\name{gimmepac}
\alias{gimmepac}
\title{gimmepac: install all previous packages after an r update}
\usage{
gimmepac(path)
}
\arguments{
\item{path}{path to which the packages were saved for your previous version\cr
do .Library to find current and then back track shold be something like this\cr
/Library/Frameworks/R.framework/Versions/}
}
\description{
fix the annoying thing where all your packages disappear after an r update
}
\examples{
gimmepac ("/Library/Frameworks/R.framework/Versions/3.4/Resources/library")

}
