% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histthem.R
\name{histthem}
\alias{histthem}
\title{serial histogram}
\usage{
histthem(dat, var)
}
\arguments{
\item{dat}{df name}

\item{var}{vector of variables}
}
\description{
winsorize to next lowest/highest value based on user input sd from the mean
}
