% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\title{LINKAGES kill function}
\usage{
kill(nspec, ntrees, slta, sltb, dbh, agemx, ksprt, sprtmn, sprtmx, iage, nogro,
  tl, rtst, fwt, max.ind, frt, ncohrt)
}
\arguments{
\item{nspec}{number of species}

\item{ntrees}{number of trees of each species}

\item{slta}{parameter to calculate crown area from diameter}

\item{sltb}{parameter to calculate crown area from diameter}

\item{dbh}{diameter of each individual}

\item{agemx}{max age}

\item{ksprt}{flags if stump can sprout}

\item{sprtmn}{minimum diameter for a stump to sprout}

\item{sprtmx}{maximum diameter for a stump to sprout}

\item{iage}{age of each individual}

\item{nogro}{flags slow growing individuals}

\item{tl}{leaf litter quality class}

\item{rtst}{root-shoot ratio for each species}

\item{fwt}{leaf weight per unit crown area}

\item{max.ind}{maximum number of individuals}

\item{frt}{foliage retention time}

\item{ncohrt}{number of cohorts}
}
\value{
dbh diameter of each individual

ntrees number of trees of each species

iage age of each individual

nogro flags individuals growing slowly

ncohrt number of cohorts

tyl total yearly litter,

ksprt flags stumps that could sprout
}
\description{
KILL KILLS TREES BY AGE DEPENDENT MORTALITY (ONLY 1%
  REACH MAXIMUM AGE) AND AGE INDEPENDENT MORTALITY (PROBABILITY OF
  SURVIVING 10 CONSECUTIVE YEARS OF SLOW GROWTH (SEE GROW) = 1%).
  DECISIONS ON WHETHER OR NOT TO KILL A TREE ARE PARTLY BASED ON
  RANDOM NUMBERS SUPPLIED BY rand.
  KILL ALSO CALCULATES LITTER AMOUNTS, WHICH ARE DECAYED IN
  SUBROUTINE DECOMP.
}
\author{
Ann Raiho
}

