% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergethem.R
\name{mergethem}
\alias{mergethem}
\title{merge multiple datasets together}
\usage{
mergethem(...)
}
\arguments{
\item{...}{list of dataframes. the matching column must be first variable}
}
\description{
merge multiple datasets together
}
