% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myggcorrplot.R
\name{myggcorrplot}
\alias{myggcorrplot}
\title{myggcorrplot}
\usage{
myggcorrplot(
  corr,
  method = c("square", "circle"),
  type = c("full", "lower", "upper"),
  ggtheme = ggplot2::theme_minimal,
  title = "",
  show.legend = TRUE,
  legend.title = "Corr",
  show.diag = FALSE,
  colors = c("blue", "white", "red"),
  outline.color = "gray",
  hc.order = FALSE,
  hc.method = "complete",
  lab = FALSE,
  lab_col = "black",
  lab_size = 4,
  p.mat = NULL,
  sig.level = 0.05,
  insig = c("pch", "blank"),
  pch = 4,
  pch.col = "black",
  pch.cex = 5,
  tl.cex = 12,
  tl.col = "black",
  tl.srt = 45,
  lab.notsig = ""
)
}
\description{
merge multiple dfs together, identifier must be first column, doesn't need to be same name
}
