% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{LINKAGES output function}
\usage{
output(availn, tyln, nspec, frt, iage, slta, sltb, dbh, fwt, tyl, max.ind,
  ntrees, awp)
}
\arguments{
\item{availn}{available nitrogen}

\item{tyln}{leaf litter N content from "decomp.r"}

\item{nspec}{number of species}

\item{frt,slta,sltb,fwt}{species specific parameters}

\item{iage}{age of each individual}

\item{dbh}{diameter of each individual}

\item{tyl}{total yearly litter}

\item{max.ind}{maximum number of individual}

\item{ntrees}{number of trees of each species}

\item{awp}{aboveground woody production}
}
\value{
atot total number of trees per ha

tbar total aboveground biomass

tyln leaf litter N content

tynap total aboveground production

availn available N

bar specieis biomass t/ha
}
\description{
OUTPUT coverts model variables into ecosystem variables of interest
}
\author{
Ann Raiho
}

