% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity.R
\name{proximity}
\alias{proximity}
\title{Proximity}
\usage{
proximity(RCAmat)
}
\arguments{
\item{RCAmat}{An economy-by-product matrix with the RCA values.}
}
\value{
A product-by-product matrix with the proximity values.
}
\description{
Compute the proximity \insertCite{hidalgo2007product}{pecan} based on the RCA
values.
}
\references{
\insertAllCited{}
}
