% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{table 1
tired of making table 1's}
\usage{
table1(df)
}
\arguments{
\item{df}{dataframe; containing numeric and/or logical columns to calculate correlations for}
}
\value{
a correlation matrix
}
\description{
table 1
tired of making table 1's
}
