% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tempe.R
\name{tempe}
\alias{tempe}
\title{LINKAGES tempe function}
\usage{
tempe(temp.vec)
}
\arguments{
\item{temp.vec}{mean temperature for each month}
}
\value{
degd total growing degree days for each year
}
\description{
Calculates total growing degree days for each year, i
}
\author{
Ann Raiho
}

