% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{winsorize function}
\usage{
winsor(x, sd)
}
\arguments{
\item{x}{vector of data}

\item{sd}{the number of standard deviations from mean as the criterion for winsorizing}
}
\description{
winsorize to next lowest/highest value based on user input sd from the mean
}
