% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergethem.R
\name{mergethem}
\alias{mergethem}
\title{mergethem}
\usage{
mergethem(...)
}
\arguments{
\item{...}{dataframes separated by comma}
}
\description{
merge multiple dfs together, identifier must be first column, doesn't need to be same name
}
